Index all content at category level and omit removed APIs from release notes
