+++ src/content/apiDocs/appeals/appealsReleaseNotes.mdx
@@ -1,0 +1,2 @@
+### Month 99, 2099
+(v1) These are dummy release notes.  Link to PRs here!
