+++ src/content/apiDocs/benefits/loanGuarantyReleaseNotes.mdx
@@ -1,3 +1,3 @@
-### March 10, 2020 (is this the right date???)
+### March 10, 2020
 
 Launch of the Loan Guaranty API
