+++ src/content/apiDocs/vaForms/index.ts
@@ -1,0 +1,16 @@
+import { IApiCategoryContent } from '../../../apiDefs/schema';
+import VAFormsIntro from './vaFormsIntro.mdx';
+import VAFormsOverview from './vaFormsOverview.mdx';
+import VAFormsReleaseNotes from './vaFormsReleaseNotes.mdx';
+
+const vaFormsContent: IApiCategoryContent = {
+  intro: VAFormsIntro,
+  overview: VAFormsOverview,
+  placardText: 'Look up VA forms and check for new versions',
+  shortDescription: 'Look up VA forms and check for new versions.',
+};
+
+export {
+  vaFormsContent,
+  VAFormsReleaseNotes,
+};
